Cleaned up and reorganized the javaScript files to better represent the client/server relationship
