--- main.js
@@ -1,19 +1,0 @@
-/**
- * July 11th, 2018
- * Chess App Project for CodeSchool 2018
- * Programmed by
- * Dockstader, Patrick
- * Richardson, Makalee
- * Wright, Grady MK
- * With instruction from
- * Eames, Andrew
- * Holt, DJ
- * Kearl, Matt
- * Wilkinson, Kaden
- */
-/**
- * main.js
- * This javaScript file is where the game state and "flow" is instantiated.
- * We use a game object to check for important states such as check, checkmate, and stalemate.
- */
-
